Move LambdaJar to core module
