+++ java/cdk/src/main/java/sleeper/cdk/stack/PropertiesStack.java
@@ -26,8 +26,8 @@ import software.constructs.Construct;
 
 import sleeper.cdk.jars.BuiltJars;
 import sleeper.cdk.jars.LambdaCode;
-import sleeper.cdk.jars.LambdaJar;
 import sleeper.cdk.util.Utils;
+import sleeper.core.deploy.LambdaJar;
 import sleeper.core.properties.instance.InstanceProperties;
 
 import java.util.HashMap;
