+++ java/cdk/src/main/java/sleeper/cdk/stack/bulkimport/EksBulkImportStack.java
@@ -63,9 +63,9 @@ import software.constructs.Construct;
 
 import sleeper.cdk.jars.BuiltJars;
 import sleeper.cdk.jars.LambdaCode;
-import sleeper.cdk.jars.LambdaJar;
 import sleeper.cdk.stack.CoreStacks;
 import sleeper.cdk.util.Utils;
+import sleeper.core.deploy.LambdaJar;
 import sleeper.core.properties.instance.CdkDefinedInstanceProperty;
 import sleeper.core.properties.instance.InstanceProperties;
