+++ java/core/src/main/java/sleeper/core/deploy/LambdaJar.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.cdk.jars;
+package sleeper.core.deploy;
 
 import sleeper.core.SleeperVersion;
