+++ java/core/src/test/java/sleeper/core/deploy/LambdaJarTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.cdk.jars;
+package sleeper.core.deploy;
 
 import org.junit.jupiter.api.Test;
